Перенос APP_DEFAULT_AVATAR и BsStyleTypes в constants/markup
