+++ src/constants/markup.js
@@ -1,3 +1,5 @@
+export const APP_DEFAULT_AVATAR = 'img/icon-avatar.svg';
+
 export const AppImgShapes = {
   USER_AVATAR: {
     width: 32,
@@ -7,4 +9,15 @@ export const AppImgShapes = {
 
 export const AppThemeClasses = {
   HEADER_DARK: 'bg-light border-bottom',
+};
+
+export const BsStyleTypes = {
+  PRIMARY: 'primary',
+  SECONDARY: 'secondary',
+  SUCCESS: 'success',
+  DANGER: 'danger',
+  WARNING: 'warning',
+  INFO: 'info',
+  LIGHT: 'light',
+  DARK: 'dark',
 };
